+++ src/Autocomplete/src/Form/AutocompleteChoiceTypeExtension.php
@@ -160,11 +160,6 @@ final class AutocompleteChoiceTypeExtension extends AbstractTypeExtension
             'extra_options' => [],
         ]);
 
-        // if autocomplete_url is passed, then HTML options are already supported
-        $resolver->setNormalizer('options_as_html', static function (Options $options, $value) {
-            return null === $options['autocomplete_url'] ? $value : false;
-        });
-
         $resolver->setNormalizer('preload', static function (Options $options, $value) {
             if (\is_bool($value)) {
                 $value = $value ? 'true' : 'false';