+++ src/LiveComponent/src/Util/ChildComponentPartialRenderer.php
@@ -12,6 +12,7 @@
 namespace Symfony\UX\LiveComponent\Util;
 
 use Psr\Container\ContainerInterface;
+use Symfony\Component\HttpKernel\Exception\BadRequestHttpException;
 use Symfony\Contracts\Service\ServiceSubscriberInterface;
 use Symfony\UX\LiveComponent\LiveComponentHydrator;
 use Symfony\UX\LiveComponent\Metadata\LiveComponentMetadataFactory;
@@ -27,6 +28,8 @@ use Twig\Runtime\EscaperRuntime;
  */
 class ChildComponentPartialRenderer implements ServiceSubscriberInterface
 {
+    private const VALID_TAG = '/\A[a-zA-Z][a-zA-Z0-9-]*+\z/';
+
     public function __construct(
         private FingerprintCalculator $fingerprintCalculator,
         private TwigAttributeHelperFactory $attributeHelperFactory,
@@ -69,7 +72,11 @@ class ChildComponentPartialRenderer implements ServiceSubscriberInterface
 
         // only send back the props that are allowed to be updated from the parent
         $readonlyDehydratedProps = $liveMetadata->getOnlyPropsThatAcceptUpdatesFromParent($props);
-        $readonlyDehydratedProps = $this->getLiveComponentHydrator()->addChecksumToData($readonlyDehydratedProps);
+        $readonlyDehydratedProps = $this->getLiveComponentHydrator()->addChecksumToData(
+            $readonlyDehydratedProps,
+            $componentName,
+            LiveComponentHydrator::CHECKSUM_SLOT_PROPS_FROM_PARENT,
+        );
 
         $attributesCollection->setPropsUpdatedFromParent($readonlyDehydratedProps);
         $attributes = $attributesCollection->toArray();
@@ -86,6 +93,10 @@ class ChildComponentPartialRenderer implements ServiceSubscriberInterface
      */
     private function createHtml(array $attributes, string $childTag): string
     {
+        if (!preg_match(self::VALID_TAG, $childTag)) {
+            throw new BadRequestHttpException('Invalid child tag.');
+        }
+
         $attributes['data-live-preserve'] = true;
         $attributes = new ComponentAttributes($attributes, $this->twig->getRuntime(EscaperRuntime::class));